refactor: pre-compute chart snapshots for all ranges in store on each recordSnapshot
